requirements.txt
absl-py==0.12.0
asgiref==3.3.4
astunparse==1.6.3
beautifulsoup4==4.9.3
bs4==0.0.1
cachetools==4.2.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Django==3.2.2
djangorestframework==3.12.4
flatbuffers==1.12
gast==0.4.0
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==2.10
joblib==1.0.1
Keras==2.4.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
lxml==4.6.3
Markdown==3.3.4
nltk==3.6.2
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
Pillow==8.2.0
protobuf==3.16.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.6.3
selenium==3.141.0
six==1.15.0
sklearn==0.0
soupsieve==2.2.1
sqlparse==0.4.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0rc3
tensorflow-estimator==2.5.0rc0
termcolor==1.1.0
threadpoolctl==2.1.0
torch==1.8.1
tqdm==4.60.0
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
